[BACK-1182] Change to Unix timestamps
To align with other Pocket API in active development, we need to serve timestamps
not as ISO strings but as Unix timestamps as well. This will standardise returned
values across our various subgraphs and also simplify generating data for Snowplow
events.

- Fixed the issue with `createdAt` and `updatedAt` values not matching - precision
on fractions of a second was different.

- Added a UnixTimestamp field resolver for timestamp values. The New Tab scheduled
date remains a date - only those timestamps that have a time component have been
converted to the new format.

- Updated tests.