feat(types): [BACK-1143] Create GraphQL enums for topic, rejection reason types
Another change where we only add an enum to the admin GraphQL schema so that
it's not hardcoded on the frontend.

The topics enum is added to be used in creating/editing/filtering corpus items,
and the rejection reasons enum is for being able to reject a prospect and
specify one or more reasons for doing so on the frontend.