chore(tests): Improve flaky Snowplow test suite.
The test suite for events sent on `reviewed_corpus_item` events was flaky -
there were two tests and on occasion, both locally and in CircleCI, the tests
would fail as Snowplow has a delay (that we cater for already) in returning
the good events.

Typically, this would mean the first test would see zero events when three were
expected and then the second test would see four good events when just the one
(for a rejected corpus item) was expected. Now these are combined into a single
test that won't have this problem.