[BACK-1219] Fix: error on not being able to delete an item doesn't appear (#294)
It appears that once we switched to the error handler from the `apollo-utils`
package throwing any old JavaScript error from the resolvers is not enough,
it has to be one of the Apollo Server errors or it won't make its way all
the way to the GraphQL server.

Now throwing an ApolloError where a JS Error was thrown before. This is
the only error affected by this change - everywhere else the code is already
throwing Apollo Server errors or descendants thereof from the `apollo-utils`
package.