fix: [BACK-1360] Fix indecipherable error messages
The error handler from `apollo-utils` masks errors that we want to
expose to end users (aka the Curation Tools Frontend) of the admin
subgraph. Removing the error handler appears to be the most sensible
option.

Note that integration tests were and are passing unchanged as they
use a separate admin server bootstrap script without the Sentry plugin
and that script was never updated to use the error handler, so the bug
was not discovered until testing the frontend functionality made it
apparent.