feat(alarms): [BACK-1617] Create alarms for Curated Corpus API
Following the release of Unified Home, we need to set up at least
a couple of alarms as a baseline to make sure we're monitoring this
service. A substantial review of this service's SLA and alarms is coming
as part of ticket 1608 - this PR is implementing two of the proposed alarms
from the WIP SLA page referred to in that ticket.